+++ src/system/path.ts
@@ -1,11 +1,9 @@
 import { isAbsolute as _isAbsolute, basename } from 'path';
 import { isLinux, isWindows } from '@env/platform';
-// TODO@eamodio don't import from string here since it will break the tests because of ESM dependencies
-// import { CharCode } from './string';
 
 export { basename, dirname, extname, join as joinPaths } from 'path';
 
-const slash = 47; //slash;
+const slash = 47; //CharCode.Slash;
 
 const driveLetterNormalizeRegex = /(?<=^\/?)([A-Z])(?=:\/)/;
 const hasSchemeRegex = /^([a-zA-Z][\w+.-]+):/;