+++ src/test/suite/index.ts
@@ -1,30 +1,40 @@
-import * as path from 'path';
+import path from 'path';
 import { glob } from 'glob';
 import Mocha from 'mocha';
 
-export async function run(): Promise<void> {
+export function run(): Promise<void> {
 	// Create the mocha test
 	const mocha = new Mocha({
-		ui: 'bdd',
+		ui: 'tdd',
 		color: true,
 	});
 
 	const testsRoot = path.resolve(__dirname, '..');
 
-	const files = await glob('**/**.test.js', { cwd: testsRoot });
+	return new Promise((c, e) => {
+		glob('**/**.test.js', { cwd: testsRoot })
+			.then(files => {
+				// Add files to the test suite
+				files.forEach(f => {
+					mocha.addFile(path.resolve(testsRoot, f));
+				});
 
-	// Add files to the test suite
-	files.forEach(f => mocha.addFile(path.resolve(testsRoot, f)));
-
-	try {
-		// Run the mocha test
-		mocha.run(failures => {
-			if (failures > 0) {
-				throw new Error(`${failures} tests failed.`);
-			}
-		});
-	} catch (ex) {
-		console.error(ex);
-		throw ex;
-	}
+				try {
+					// Run the mocha test
+					mocha.run(failures => {
+						if (failures > 0) {
+							e(new Error(`${failures} tests failed.`));
+						} else {
+							c();
+						}
+					});
+				} catch (err) {
+					console.error(err);
+				}
+			})
+			.catch((err: unknown) => {
+				console.error(err);
+				return err;
+			});
+	});
 }